Some enhancements
Used struct for value objects
Used Decimal for amount
Managed migration and seeding at startup
Added a middleware to catch exceptions

+++ Spending.Infrastructure/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -56,7 +56,7 @@ namespace Spending.Infrastructure.Migrations
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<double>("Amount");
+                    b.Property<decimal>("Amount");
 
                     b.Property<string>("Comment")
                         .IsRequired();
