+++ Spending/ViewModels/CreateSpendingViewModel.cs
@@ -1,6 +1,7 @@
 ﻿namespace Spending.ViewModels
 {
     using Spending.Domain.Entity;
+    using Spending.Domain.Enum;
     using System;
 
     public class CreateSpendingViewModel
@@ -8,7 +9,7 @@
         public long SpenderId { get; set; }
         public DateTime? Date { get; set; }
         public SpendingNature Nature { get; set; }
-        public double? Amount { get; set; }
+        public decimal? Amount { get; set; }
         public long CurrencyId { get; set; }
         public string Comment { get; set; }
     }
